Fix assignment of state_val/data_val
Variable declaration assignments such as

	reg foo = 1;

are interpreted by some tools as

	reg foo;
	always @(*) foo = 1;

and not as

	reg foo;
	initial foo = 1;

This can cause simulation-synthesis mismatches, as the resultant value
of foo depends on how the tool is feeling.

Variable declaration assignments are used in four places. However, the
variables in question are subsequently assigned in their entirety,
Remove these assignments.

Closes: #4